Add files via upload
Created file for simulating the equilibrium concentration of unbound positive holes as a function of depth. This file is in an early stage. No graphs have been set up, and the solutions to the partial differential equation have not been calculated yet. I'm still trying to figure out how to solve specific PDEs in MATLAB.